refactor(server): extract wire-level types into server/wire
Carves the small CP↔worker RPC types and PG cancel-request helpers out
of server/ into a new server/wire subpackage with no dependency on
github.com/duckdb/duckdb-go.

Symbols moved:
  - BackendKey (from server/server.go)
  - GenerateSecretKey + the previously private generateSecretKey
    (from server/exports.go and server/conn.go)
  - WorkerControlMetadata (from server/worker_control.go)
  - WorkerActivationPayload (from server/worker_activation.go), with its
    DuckLake field now typed as ducklake.Config directly rather than via
    the server.DuckLakeConfig alias
  - WorkerCreateSessionPayload, WorkerDestroySessionPayload,
    WorkerHealthCheckPayload (from server/worker_control.go)

Backward compatibility preserved via type aliases in
server/worker_control.go and server/worker_activation.go (now thin
alias files), plus a `var GenerateSecretKey = wire.GenerateSecretKey`
re-export in server/exports.go and a `type BackendKey = wire.BackendKey`
alias in server/server.go.

  go list -deps ./server/wire | grep duckdb-go   # empty

The control plane uses many of these symbols via the existing
server.X spellings; those continue to compile via the aliases. New
code (and the eventual cmd/duckgres-controlplane binary) should
import server/wire directly.

Also drops two now-unused imports from server/conn.go (crypto/rand
and math/big — both were only used by the moved generateSecretKey).

Verified:
  - go build ./... clean
  - go build -tags kubernetes ./... clean
  - go test -short ./server/wire/... ./server/... ./controlplane/...
    ./duckdbservice/... — all green
  - go list -deps ./server/wire | grep duckdb-go is empty
